LPS-70229 Distribute the logic in different modules and rename source foramtting
