+++ adaptive-media-asset-publisher-web/build.gradle
@@ -7,6 +7,7 @@ dependencies {
 	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
 	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	provided project(":apps:adaptive-media:adaptive-media-api")
+	provided project(":apps:adaptive-media:adaptive-media-content-transformer")
+	provided project(":apps:adaptive-media:adaptive-media-image-content-transformer")
 	provided project(":apps:adaptive-media:adaptive-media-web")
 }
