+++ adaptive-media-blogs-web/src/main/java/com/liferay/adaptive/media/blogs/web/internal/filter/BlogsPortletFilter.java
@@ -14,8 +14,8 @@
 
 package com.liferay.adaptive.media.blogs.web.internal.filter;
 
-import com.liferay.adaptive.media.processor.content.ContentProcessorHandler;
-import com.liferay.adaptive.media.web.constants.ContentTypes;
+import com.liferay.adaptive.media.content.transformer.ContentTransformerHandler;
+import com.liferay.adaptive.media.image.content.transformer.constants.ContentTransformerContentTypes;
 import com.liferay.blogs.web.constants.BlogsPortletKeys;
 import com.liferay.portal.kernel.servlet.BufferCacheServletResponse;
 import com.liferay.portal.kernel.servlet.ServletResponseUtil;
@@ -74,10 +74,10 @@ public class BlogsPortletFilter implements RenderFilter {
 
 		String content = bufferCacheServletResponse.getString();
 
-		String processedContent = contentProcessorHandler.process(
-			ContentTypes.HTML, content);
+		String transformedContent = contentTransformerHandler.process(
+			ContentTransformerContentTypes.HTML, content);
 
-		ServletResponseUtil.write(httpServletResponse, processedContent);
+		ServletResponseUtil.write(httpServletResponse, transformedContent);
 	}
 
 	@Override
@@ -85,7 +85,7 @@ public class BlogsPortletFilter implements RenderFilter {
 	}
 
 	@Reference
-	protected ContentProcessorHandler contentProcessorHandler;
+	protected ContentTransformerHandler contentTransformerHandler;
 
 	@Reference
 	protected Portal portal;
