+++ adaptive-media-content-transformer/build.gradle
@@ -1,0 +1,16 @@
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
+
+dependencies {
+	provided group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
+	provided group: "com.liferay", name: "com.liferay.osgi.service.tracker.collections", version: "2.0.0"
+	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
+	provided group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
+	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":apps:adaptive-media:adaptive-media-api");
+
+	testCompile group: "junit", name: "junit", version: "4.12"
+	testCompile group: "org.apache.felix", name: "org.apache.felix.framework", version: "5.4.0"
+	testCompile group: "org.mockito", name: "mockito-core", version: "1.10.8", transitive: true
+}
