+++ adaptive-media-image-content-transformer/build.gradle
@@ -1,0 +1,13 @@
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
+
+dependencies {
+	provided group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
+	provided group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
+	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+	provided project(":apps:adaptive-media:adaptive-media-api");
+	provided project(":apps:adaptive-media:adaptive-media-content-transformer");
+	provided project(":apps:adaptive-media:adaptive-media-image-api");
+	provided project(":apps:adaptive-media:adaptive-media-image-impl");
+	provided project(":apps:adaptive-media:adaptive-media-web");
+}
