+++ adaptive-media-image-content-transformer/src/main/java/com/liferay/adaptive/media/image/content/transformer/constants/ContentTransformerContentTypes.java
@@ -1,0 +1,27 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ * <p>
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ * <p>
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.adaptive.media.image.content.transformer.constants;
+
+import com.liferay.adaptive.media.content.transformer.ContentTransformerContentType;
+
+/**
+ * @author Alejandro Tardín
+ */
+public class ContentTransformerContentTypes {
+
+	public static final ContentTransformerContentType<String> HTML =
+		new ContentTransformerContentType<String>() {};
+
+}
