+++ adaptive-media-image-web/build.gradle
@@ -35,6 +35,7 @@ dependencies {
 	provided group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:adaptive-media:adaptive-media-api")
+	provided project(":apps:adaptive-media:adaptive-media-content-transformer")
 	provided project(":apps:adaptive-media:adaptive-media-image-impl")
 	provided project(":apps:adaptive-media:adaptive-media-web")
