--- adaptive-media-web/src/main/java/com/liferay/adaptive/media/web/constants/ContentTypes.java
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- * <p>
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- * <p>
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.adaptive.media.web.constants;
-
-import com.liferay.adaptive.media.processor.content.ContentType;
-
-/**
- * @author Alejandro Tardín
- */
-public class ContentTypes {
-
-	public static final ContentType<String> HTML = new ContentType<String>() {};
-
-}
